Started working on the documentation (wip) Moved sequencer back into private, and had to write a function runprivate2, that can handle two cellarray as output - there probably is a smarter solution to this)
New Feature: In Exp.Dim it is now possible to vary the frequency of a frequency swept pulse with linear and non linear increments. On top of that it is now also possible to vary pulse parameters that require a vector (e.g. Frequency , A0 or n) by writing 'p1.A0(3)', etc. with linear and nonlinear increments. This should give a lot more flexibility in varying pulse parameters

--HG--
branch : evolve